carry the editor's command payload in a uintptr_t
ICommandHandler::HandleCommand takes a command id and a dwData, and dwData is
almost always a pointer: the object set a state wants filled in, the view or
command handler a dialog is asking for, the SEditParameters a window reads back
out of its state, a std::string, a CTPoint. The parameter was 32 bits wide.

On x86 that worked, which is why it was written this way and why the sweep that
put the cstdint types in kept it: DWORD and a pointer were both four bytes. On
x64 every one of those calls truncates the pointer on the way in, and the
handler on the other side rebuilds a pointer out of half an address and reads
it. It compiles, with a warning per site, and it crashes as soon as anything
sits above 4 GB, which on a 64-bit Windows process is most of the heap.

So the payload becomes uintptr_t, which is pointer-sized by definition on every
target, and the casts at both ends follow. 273 substitutions across 158 files,
all mechanical: uint32_t dwData in the three interface declarations and their
168 overrides, reinterpret_cast<uint32_t> at the 97 call sites that put a
pointer in, and ten C style casts in ScriptCameraState.cpp that do the same
thing to a float * or an int *. The receiving casts needed no edit at all; they
stopped truncating the moment the source stopped being 32 bits.

DIDEVICEOBJECTDATA::dwData in Input and port/dinput.h is a different dwData,
genuinely 32 bits, and is left alone.

This is 143 of the tree's 166 pointer truncation warnings. The rest are
Scintilla's message interface and two Win32 handles, which are a different
mechanism and a separate change.

+++ Versions/Temporary/Engine/Sources/ED_B2/ChapterState.cpp
@@ -33,7 +33,7 @@ CChapterState::CChapterState( CChapterEditor *_pChapterEditor ) :
 
 void CChapterState::Enter()
 {
-//	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_ENABLE_GAME_INPUT, reinterpret_cast<uint32_t>( new CChapterInterfaceCommand( new CChapterInterface() ) ) );
+//	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_ENABLE_GAME_INPUT, reinterpret_cast<uintptr_t>( new CChapterInterfaceCommand( new CChapterInterface() ) ) );
 
 	// clear the scene
 	Singleton<IUIScene>()->Create();
@@ -545,7 +545,7 @@ IView* CChapterState::ClearView()
 	IView *pView = 0;
 	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_PC_DIALOG, 
 																												ID_PC_DIALOG_GET_VIEW, 
-																												reinterpret_cast<uint32_t>(&pView) );
+																												reinterpret_cast<uintptr_t>(&pView) );
 	if ( pView != 0 )
 	{
 		pView->RemoveViewManipulator();
@@ -559,7 +559,7 @@ void CChapterState::SetView( IView *pView )
 {
 	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_PC_DIALOG, 
 																												ID_PC_DIALOG_GET_VIEW, 
-																												reinterpret_cast<uint32_t>(&pView) );
+																												reinterpret_cast<uintptr_t>(&pView) );
 	bool bNeedCreateTree = true;
 	if ( pView != 0 )
 	{
@@ -573,7 +573,7 @@ void CChapterState::SetView( IView *pView )
 	ICommandHandler *pCommandHandler = 0;
 	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_PC_DIALOG, 
 																												ID_PC_DIALOG_GET_COMMAND_HANDLER, 
-																												reinterpret_cast<uint32_t>(&pCommandHandler) );
+																												reinterpret_cast<uintptr_t>(&pCommandHandler) );
 	if ( pCommandHandler != 0 )
 	{
 		pCommandHandler->HandleCommand( ID_PC_EXPAND_ALL, 0 );
@@ -587,7 +587,7 @@ void CChapterState::ClearMaskManipulator()
 		IView *pView = 0;
 		Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_PC_DIALOG,
 																													ID_PC_DIALOG_GET_VIEW, 
-																													reinterpret_cast<uint32_t>(&pView) );
+																													reinterpret_cast<uintptr_t>(&pView) );
 		if ( pView != 0 )
 		{
 			pView->RemoveViewManipulator();
