+++ Versions/Temporary/Engine/Sources/ED_B2_M1/AIGeneralState.h
@@ -43,7 +43,7 @@ public:
 	void OnKeyDown( unsigned nChar, unsigned nRepCnt, unsigned nFlags );
 
 	// ICommandHandler
-	virtual bool HandleCommand( unsigned nCommandID, uint32_t dwData );
+	virtual bool HandleCommand( unsigned nCommandID, uintptr_t dwData );
 	virtual bool UpdateCommand( unsigned nCommandID, bool *pbEnable, bool *pbCheck );
 
 	void GetDataFromDB();
