+++ Versions/Temporary/Engine/Sources/ED_B2_M1/AIGeneralWindow.h
@@ -117,7 +117,7 @@ public:
 	void OnCancel() {}
 
 	// ICommandHandler
-	virtual bool HandleCommand( unsigned nCommandID, uint32_t dwData );
+	virtual bool HandleCommand( unsigned nCommandID, uintptr_t dwData );
 	virtual bool UpdateCommand( unsigned nCommandID, bool *pbEnable, bool *pbCheck );
 	virtual void NotifyHandler();
 	void NotifyHandler( SAIGeneralPointsWindowData::EAIGenPointsLastAction eAction );
