+++ Versions/Temporary/Engine/Sources/ED_B2_M1/BridgeState.cpp
@@ -29,7 +29,7 @@ const int CBridgeState::TERMINATOR_COUNT = 2;
 bool CBridgeState::CanAddBridge()
 {
 	SObjectSet objectSet;
-	if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uint32_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
+	if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uintptr_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
 	{
 		return ( objectSet.szObjectTypeName == "BridgeRPGStats" ) &&
 					 ( !objectSet.objectNameSet.empty() );
@@ -44,7 +44,7 @@ void CBridgeState::InsertObjectEnter()
 	ClearData();
 	//
 	CWaitCursor waitCursor;
-	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uint32_t>( &objectSet ) );
+	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uintptr_t>( &objectSet ) );
 }
 
 
@@ -254,7 +254,7 @@ void CBridgeState::InsertBridge( SBridgeInfo::EDirection direction, bool bFixSta
 		// получаем размер 
 		if ( objectSet.szObjectTypeName.empty() )
 		{
-			if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uint32_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
+			if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uintptr_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
 			{
 				CPtr<IManipulator> pBridgeRPGStatsManipulator = pResourceManager->CreateObjectManipulator( objectSet.szObjectTypeName, objectSet.objectNameSet.begin()->first );
 				if ( pBridgeRPGStatsManipulator )
