+++ Versions/Temporary/Engine/Sources/ED_B2_M1/BuildingEditor.cpp
@@ -177,7 +177,7 @@ void CBuildingEditor::Destroy()
 	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_CLEAR, 0 );
 }
 
-bool CBuildingEditor::HandleCommand( unsigned nCommandID, uint32_t dwData )
+bool CBuildingEditor::HandleCommand( unsigned nCommandID, uintptr_t dwData )
 {
 	return false;
 }
