+++ Versions/Temporary/Engine/Sources/ED_B2_M1/EditorTestState.cpp
@@ -19,7 +19,7 @@ void CEditorTestState::Enter()
 {
 	NI_ASSERT( pEditorTestEditor != 0, "CEditorTestState::Enter(), pEditorTestEditor == 0" );
 	// Turn On Game Input
-	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_ENABLE_GAME_INPUT, reinterpret_cast<uint32_t>( new CEditorTestInterfaceCommand( new CEditorTestInterface() ) ) );
+	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_ENABLE_GAME_INPUT, reinterpret_cast<uintptr_t>( new CEditorTestInterfaceCommand( new CEditorTestInterface() ) ) );
 	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_UPDATE, 0 );
 }
