+++ Versions/Temporary/Engine/Sources/ED_B2_M1/EntrenchmentState.cpp
@@ -166,7 +166,7 @@ void CEntrenchmentState::InsertEntrenchment()
 }
 
 
-bool CEntrenchmentState::HandleCommand( unsigned nCommandID, uint32_t dwData )
+bool CEntrenchmentState::HandleCommand( unsigned nCommandID, uintptr_t dwData )
 {
 	if ( CMapObjectState::HandleCommand( nCommandID, dwData ) )
 	{
@@ -216,7 +216,7 @@ void CEntrenchmentState::RefreshSelectedEntrenchmentInfo()
 	//
 	SObjectSet objectSet;
 	if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, 
-		ID_OS_GET_OBJECTSET, reinterpret_cast<uint32_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
+		ID_OS_GET_OBJECTSET, reinterpret_cast<uintptr_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
 	{
 		bool bRes = ( objectSet.szObjectTypeName == "EntrenchmentRPGStats" ) && ( !objectSet.objectNameSet.empty() );
 		if ( !bRes )
