+++ Versions/Temporary/Engine/Sources/ED_B2_M1/MapInfoEditor.cpp
@@ -646,7 +646,7 @@ void CMapInfoEditor::Save( bool bSaveChanges )
 }
 
 
-bool CMapInfoEditor::HandleCommand( unsigned nCommandID, uint32_t dwData )
+bool CMapInfoEditor::HandleCommand( unsigned nCommandID, uintptr_t dwData )
 {
 	switch( nCommandID ) 
 	{
