+++ Versions/Temporary/Engine/Sources/ED_B2_M1/MapInfoState.cpp
@@ -393,7 +393,7 @@ void CMapInfoState::Enter()
 	NProgress::IteratePosition(); // 2
 	NI_ASSERT( pMapInfoEditor != 0, "CMapInfoState::Enter(), pMapInfoEditor == 0" );
 	//
-	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_ENABLE_GAME_INPUT, reinterpret_cast<uint32_t>( new CMapInfoInterfaceCommand( new CMapInfoInterface( this ) ) ) );
+	Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_SCENE, ID_SCENE_ENABLE_GAME_INPUT, reinterpret_cast<uintptr_t>( new CMapInfoInterfaceCommand( new CMapInfoInterface( this ) ) ) );
 	//
 	pMapInfoEditor->ClearMapInfoData();
 	NProgress::IteratePosition(); // 3
@@ -694,7 +694,7 @@ void CMapInfoState::Leave()
 }
 
 
-bool CMapInfoState::HandleCommand( unsigned nCommandID, uint32_t dwData )
+bool CMapInfoState::HandleCommand( unsigned nCommandID, uintptr_t dwData )
 {
 	switch( nCommandID )
 	{
