+++ Versions/Temporary/Engine/Sources/ED_B2_M1/MapInfoState.h
@@ -100,7 +100,7 @@ public:
 
 	void CancelSelection();
 	//ICommandHandler
-	bool HandleCommand( unsigned nCommandID, uint32_t dwData );
+	bool HandleCommand( unsigned nCommandID, uintptr_t dwData );
 	bool UpdateCommand( unsigned nCommandID, bool *pbEnable, bool *pbCheck );
 };
